+++ rpXsuite/RPxSuite.py
@@ -75,7 +75,8 @@ def main(args):
 
     ## Cluster protein hits
     print("Running VSEARCH")
-    cmd = ["vsearch", "--cluster_fast", args.output + "all.hits", "--id", str(args.id), "--centroids", args.output + "centroids.fasta", "--maxrejects", "0", "--threads", "6", "--uc", args.output + "clusters.txt"]
+    cmd = ["vsearch", "--cluster_fast", args.output + "all.hits", "--id", str(args.id), "--centroids", args.output + "centroids.fasta", "--uc", args.output + "clusters.txt"]
+    print(' '.join(cmd))
     process = subprocess.Popen(cmd, stdout=DEVNULL).wait()
 
     ## Return cluster information
@@ -139,6 +140,35 @@ def parse_usearch_clustering(loc):
 
     return pd.DataFrame(table)
 
+def load_genes(describe = False):
+    '''
+    Load gene thresholds and return g2t (gene -> threshold)
+
+    If describe, print a description of the genes available
+    '''
+    # Get the thresholds and recoverability
+    Rdb = pd.read_csv(os.path.dirname(__file__) + '/helper_files/SupplementalTable_S4.2.tsv', sep='\t')
+    g2t = Rdb.set_index('gene')['overall_threshold'].to_dict()
+
+    if describe:
+        # Get the descriptions
+        Ddb = pd.read_csv(os.path.dirname(__file__) + '/helper_files/SupplementalTable_S2.2.tsv', sep='\t')
+        g2d = Ddb.sort_values('description', ascending=False).set_index('gene')['description'].to_dict()
+
+        # Add
+        Rdb['description'] = Rdb['gene'].map(g2d)
+
+        # Alter
+        Rdb = Rdb[~(Rdb['gene'] == '16S')]
+        Rdb = Rdb.rename(columns={'overall_threshold':'species_threshold'})
+        Rdb = Rdb[['gene', 'species_threshold', 'description']]
+        Rdb = Rdb[~Rdb['description'].isna()]
+
+        # Print
+        print(Rdb.to_string(index=False, max_rows=1000, line_width=200))
+
+    return(g2t)
+
 if __name__ == '__main__':
 
     parser = argparse.ArgumentParser(description= """
@@ -150,16 +180,32 @@ if __name__ == '__main__':
     # Required positional arguments
     parser.add_argument('input', nargs="+",  help="path to directories, with an asterix. eg. *.fasta, or ./assemblies/*.fasta")
     parser.add_argument("-g", "--gene", action="store", default="Ribosomal_S9", \
-        help='Marker gene to look for. Options: PGK,Ribosomal_L23,Ribosomal_L5,Ribosomal_L3,Ribosomal_L6,Ribosomal_S17,Ribosomal_S9,Ribosomal_S8,Ribosomal_S11,Ribosomal_S13,Ribosomal_L10,Ribosomal_L4,tRNA-synt_1d,GrpE,Methyltransf_5,TIGR00001,TIGR00002,TIGR00009,TIGR00012,TIGR00019,TIGR00029,TIGR00043,TIGR00059,TIGR00060,TIGR00061,TIGR00062,TIGR00064,TIGR00082,TIGR00086,TIGR00092,TIGR00115,TIGR00116,TIGR00152,TIGR00158,TIGR00165,TIGR00166,TIGR00168,TIGR00234,TIGR00337,TIGR00344,TIGR00362,TIGR00388,TIGR00389,TIGR00392,TIGR00396,TIGR00408,TIGR00409,TIGR00414,TIGR00418,TIGR00420,TIGR00422,TIGR00435,TIGR00436,TIGR00442,TIGR00459,TIGR00460,TIGR00468,TIGR00471,TIGR00472,TIGR00487,TIGR00496,TIGR00575,TIGR00631,TIGR00663,TIGR00775,TIGR00810,TIGR00855,TIGR00922,TIGR00952,TIGR00959,TIGR00963,TIGR00964,TIGR00967,TIGR00981,TIGR01009,TIGR01011,TIGR01017,TIGR01021,TIGR01024,TIGR01029,TIGR01030,TIGR01031,TIGR01032,TIGR01044,TIGR01049,TIGR01050,TIGR01059,TIGR01063,TIGR01066,TIGR01067,TIGR01071,TIGR01079,TIGR01164,TIGR01169,TIGR01171,TIGR01391,TIGR01393,TIGR01632,TIGR01953,TIGR02012,TIGR02013,TIGR02027,TIGR02191,TIGR02350,TIGR02386,TIGR02387,TIGR02397,TIGR02432,TIGR02729,TIGR03263,TIGR03594')
-    parser.add_argument("-i", "--id", action="store", default=0.97, \
-        help='Marker gene percent identity cutoff to use when clustering. ')
+        help='Marker gene to look for. Use the argument "--describe_genes" for a list of options')
+    parser.add_argument("-i", "--id", action="store", default=0, \
+        help='Marker gene percent identity cutoff to use when clustering. By default use the optimal threshold for species-level clustering')
     parser.add_argument("-o", "--output", action="store", default="./marker_genes/", \
         help='Output directory')
     parser.add_argument("-s", "--score_cutoff", action="store", default="cut_ga",  \
         help='An HMM score threshold to use - cut_ga, cut_nc, or cut_tc.')
     parser.add_argument("--amino_acid", dest='amino_acid', action='store_true', \
         help='Choose to cluster by amino_acid sequence instead of nucleotide sequence.')
+    parser.add_argument("--describe_genes", action='store_true', \
+        help='Print the gene options and exit')
 
     parser.set_defaults(amino_acid=False)
+    parser.set_defaults(describe_genes=False)
     args = parser.parse_args()
+
+    # describe genes
+    if args.describe_genes == True:
+        load_genes(describe=True)
+        sys.exit()
+
+    # Get the threshold
+    if args.id == 0:
+        g2t = load_genes()
+        thresh = g2t[args.gene]
+        args.id = thresh
+        print("ID threshold of {0} being used for gene {1}".format(args.id, args.gene))
+
     main(args)